feat(memory): share canonical recall selection across scopes
Move the existing eligibility, freshness, ranking and budget contract into the
standalone plugin bundle. Keep the control-plane facade and inspection
behavior while requiring explicit delivery scopes, including user-global.

Constraint: The installed Hermes plugin must use stdlib and bundled modules without a second ranking implementation.
Rejected: Keep a reduced provider ranker | Scope, lifecycle and budget decisions would continue to diverge.
Confidence: high
Scope-risk: moderate
Directive: Keep wildcard inspection separate from delivery and preserve reviewed store schemas.
Tested: Behavioral empty-scope RED; canonical and global-scope CLI tests; parent integration and independent phase verification; Ruff, compileall and diff checks.
Not-tested: Post-main integration and exact-head CI remain pending; relocated baseline type diagnostics are documented.
