+++ src/workflows/memory.py
@@ -27,16 +27,98 @@ from ..plugin_bundle.omh.memory_dreaming import consolidation_path as _consolida
 from ..plugin_bundle.omh.memory_governance import (
     ADMISSION_STATES,
     MEMORY_GOVERNANCE_POLICY_VERSION,
-    MEMORY_SCOPE_SCHEMA_VERSION as _V2_MEMORY_SCOPE_SCHEMA_VERSION,
-    PROJECT_MEMORY_RECORD_SCHEMA_VERSION as _V2_PROJECT_MEMORY_RECORD_SCHEMA_VERSION,
     PROJECT_MEMORY_REVIEW_RECORD_SCHEMA_VERSION as _V2_PROJECT_MEMORY_REVIEW_RECORD_SCHEMA_VERSION,
     build_retention,
     canonical_payload_digest,
     classify_memory_admission,
     contains_credential_like_material,
-    evaluate_memory_replay,
     evaluate_renderable_strings,
     stable_artifact_identity,
+)
+from ..plugin_bundle.omh.memory_recall_selector import (
+    _evaluate_memory_artifact as _shared_evaluate_memory_artifact,
+    effective_recall_configuration as effective_recall_configuration,
+    select_memory_recall,
+)
+from ..plugin_bundle.omh.memory_recall_support import (
+    LEGACY_MEMORY_SCOPE_SCHEMA_VERSION as LEGACY_MEMORY_SCOPE_SCHEMA_VERSION,
+    LEGACY_PROJECT_MEMORY_RECORD_SCHEMA_VERSION as LEGACY_PROJECT_MEMORY_RECORD_SCHEMA_VERSION,
+    MEMORY_SCOPE_SCHEMA_VERSION as MEMORY_SCOPE_SCHEMA_VERSION,
+    PROJECT_MEMORY_RECORD_SCHEMA_VERSION as PROJECT_MEMORY_RECORD_SCHEMA_VERSION,
+    _INSPECTABLE_STALE_REASONS as _INSPECTABLE_STALE_REASONS,
+    ADVISORY_FRESHNESS_REASONS as ADVISORY_FRESHNESS_REASONS,
+    DEFAULT_MEMORY_ATTENTION_TIER as DEFAULT_MEMORY_ATTENTION_TIER,
+    MAX_RETENTION_DAYS as MAX_RETENTION_DAYS,
+    MEMORY_ATTENTION_TIERS as MEMORY_ATTENTION_TIERS,
+    PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION as PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION,
+    _ADMISSION_VERACITY_DEFAULT_PCT as _ADMISSION_VERACITY_DEFAULT_PCT,
+    _ADMISSION_VERACITY_WEIGHT_PCT as _ADMISSION_VERACITY_WEIGHT_PCT,
+    _ADVISORY_NEXT_ACTIONS as _ADVISORY_NEXT_ACTIONS,
+    _AGE_TIER_BOUNDS_DAYS as _AGE_TIER_BOUNDS_DAYS,
+    _AGE_TIER_WEIGHTS as _AGE_TIER_WEIGHTS,
+    _DEFAULT_PERSPECTIVE_OBSERVER as _DEFAULT_PERSPECTIVE_OBSERVER,
+    _DUE_SOON_NEXT_ACTION as _DUE_SOON_NEXT_ACTION,
+    _EPISODE_DEFAULT_TTL_DAYS as _EPISODE_DEFAULT_TTL_DAYS,
+    _EXPIRES_SOON_DAYS as _EXPIRES_SOON_DAYS,
+    _EXPIRES_SOON_NEXT_ACTION as _EXPIRES_SOON_NEXT_ACTION,
+    _FRESHNESS_NEXT_ACTION as _FRESHNESS_NEXT_ACTION,
+    _FRESHNESS_REASON_TEXT as _FRESHNESS_REASON_TEXT,
+    _FRESHNESS_WARNING_LIMIT as _FRESHNESS_WARNING_LIMIT,
+    _MEMORY_ASCII_TOKEN as _MEMORY_ASCII_TOKEN,
+    _MEMORY_ATTENTION_RANK as _MEMORY_ATTENTION_RANK,
+    _MEMORY_CADENCE_DEFAULTS as _MEMORY_CADENCE_DEFAULTS,
+    _MEMORY_CADENCE_MAX as _MEMORY_CADENCE_MAX,
+    _MEMORY_CJK_RUN as _MEMORY_CJK_RUN,
+    _MEMORY_PINS_LIMIT as _MEMORY_PINS_LIMIT,
+    _MEMORY_SHORT_ASCII_TOKEN as _MEMORY_SHORT_ASCII_TOKEN,
+    _MEMORY_SHORT_STOPWORDS as _MEMORY_SHORT_STOPWORDS,
+    _RECALL_RRF_K as _RECALL_RRF_K,
+    _RECALL_RRF_WEIGHTS as _RECALL_RRF_WEIGHTS,
+    _REVIEW_DEFAULT_DAYS as _REVIEW_DEFAULT_DAYS,
+    _REVIEW_DUE_SOON_DAYS as _REVIEW_DUE_SOON_DAYS,
+    _SAFE_TAG as _SAFE_TAG,
+    _SOURCE_EVIDENCE_MAX_BYTES as _SOURCE_EVIDENCE_MAX_BYTES,
+    _TEMPORAL_QUERY_CUES as _TEMPORAL_QUERY_CUES,
+    _TEMPORAL_QUERY_PHRASES as _TEMPORAL_QUERY_PHRASES,
+    _age_tier as _age_tier,
+    _attach_recall_ranking as _attach_recall_ranking,
+    _attention_disclosure as _attention_disclosure,
+    _cadence_value as _cadence_value,
+    _competition_ranks as _competition_ranks,
+    _earliest_deadline as _earliest_deadline,
+    _empty_recall_pack as _empty_recall_pack,
+    _freshness_warnings as _freshness_warnings,
+    _local_source_digest as _local_source_digest,
+    _memory_recall_score as _memory_recall_score,
+    _memory_tokens as _memory_tokens,
+    _normalize_evaluator_timestamps as _normalize_evaluator_timestamps,
+    _normalize_scope as _normalize_scope,
+    _normalize_tags as _normalize_tags,
+    _parse_utc as _parse_utc,
+    _parse_utc_naive_as_utc as _parse_utc_naive_as_utc,
+    _perspective_projection as _perspective_projection,
+    _ranking_field as _ranking_field,
+    _ranking_flag as _ranking_flag,
+    _recall_evidence_fields as _recall_evidence_fields,
+    _recall_exclusion as _recall_exclusion,
+    _recall_item as _recall_item,
+    _recall_query_intent as _recall_query_intent,
+    _record_attention_tier as _record_attention_tier,
+    _record_perspective_matches as _record_perspective_matches,
+    _record_staleness as _record_staleness,
+    _redact_admitted_text as _redact_admitted_text,
+    _redact_nested_metadata as _redact_nested_metadata,
+    _redacted_metadata_label as _redacted_metadata_label,
+    _replay_evaluation as _replay_evaluation,
+    _resolve_query_intent as _resolve_query_intent,
+    _retention_class as _retention_class,
+    _scope as _scope,
+    _source_evidence_state as _source_evidence_state,
+    _string_list as _string_list,
+    _usage_bucket as _usage_bucket,
+    freshness_reason_detail as freshness_reason_detail,
+    normalize_memory_attention_tier as normalize_memory_attention_tier,
+    record_attention_tier as record_attention_tier,
 )
 from ..paths import OmhPaths, project_identity
 from ..profiles.setup import read_setup_profile
@@ -48,20 +130,15 @@ MEMORY_INSPECTION_SCHEMA_VERSION = "memory_inspection/v1"
 MEMORY_REVIEW_CARD_SCHEMA_VERSION = "memory_review_card/v1"
 HANDOFF_CONTEXT_PACK_SCHEMA_VERSION = "handoff_context_pack/v1"
 MEMORY_UPDATE_BATCH_SCHEMA_VERSION = "memory_update_batch/v1"
-MEMORY_SCOPE_SCHEMA_VERSION = _V2_MEMORY_SCOPE_SCHEMA_VERSION
-LEGACY_MEMORY_SCOPE_SCHEMA_VERSION = "omh_memory_scope/v1"
 MEMORY_INDEX_SCHEMA_VERSION = "omh_memory_index/v1"
 PROJECT_MEMORY_POLICY_SCHEMA_VERSION = "project_memory_policy/v1"
 PROJECT_MEMORY_STATUS_SCHEMA_VERSION = "project_memory_status/v1"
 PROJECT_MEMORY_CAPTURE_SCHEMA_VERSION = "project_memory_capture/v1"
 PROJECT_MEMORY_CANDIDATE_SCHEMA_VERSION = "project_memory_candidate/v1"
-PROJECT_MEMORY_RECORD_SCHEMA_VERSION = _V2_PROJECT_MEMORY_RECORD_SCHEMA_VERSION
-LEGACY_PROJECT_MEMORY_RECORD_SCHEMA_VERSION = "project_memory_record/v1"
 PROJECT_MEMORY_REVIEW_CARD_SCHEMA_VERSION = "project_memory_review_card/v1"
 PROJECT_MEMORY_REVIEW_QUEUE_SCHEMA_VERSION = "project_memory_review_queue/v1"
 PROJECT_MEMORY_REVIEW_RECORD_SCHEMA_VERSION = _V2_PROJECT_MEMORY_REVIEW_RECORD_SCHEMA_VERSION
 LEGACY_PROJECT_MEMORY_REVIEW_RECORD_SCHEMA_VERSION = "project_memory_review_record/v1"
-PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION = "project_memory_recall_pack/v1"
 MEMORY_RECALL_USAGE_SCHEMA_VERSION = "omh_memory_recall_usage/v1"
 MEMORY_LINEAGE_SCHEMA_VERSION = "omh_memory_lineage/v1"
 MEMORY_PERSPECTIVES_SCHEMA_VERSION = "omh_memory_perspectives/v1"
@@ -97,7 +174,7 @@ SOURCE_PRECEDENCE = {
     "wrapper_snapshot": 30,
 }
 ALLOWED_UPDATE_OPS = {"keep", "forget", "update", "change_scope", "dismiss_conflict"}
-ALLOWED_SCOPE_KINDS = {"project", "target", "thread", "run"}
+ALLOWED_SCOPE_KINDS = {"user-global", "project", "target", "thread", "run"}
 PROJECT_MEMORY_MODES = ("off", "review-first", "auto-safe")
 PROJECT_MEMORY_RECORD_TYPES = ("fact", "decision", "lesson", "procedure", "episode")
 MEMORY_ACTION_IDS = (
@@ -114,7 +191,6 @@ _SAFE_REF = re.compile(r"^[A-Za-z0-9_.:-]{1,120}$")
 # words. Running them through the ASCII-only _SAFE_REF silently dropped every
 # Korean/Japanese/Chinese tag at capture time, which meant tag scoring could
 # never fire for records written by CJK-speaking projects.
-_SAFE_TAG = re.compile(r"^[\w.:/-]{1,120}$", re.UNICODE)
 _PROMPTISH_KEYS = {"message", "prompt", "raw", "text", "body", "content", "prompt_template"}
 _PROJECT_MEMORY_RECORD_KEYS = {
     "schema_version",
@@ -214,8 +290,6 @@ _RECALL_RANKING_KEYS = {
 # both classes passed the same admission gates -- and it lands in the
 # decayed score, never in eligibility. An unknown mode fails CLOSED to the
 # lower weight: a trust signal must never default to maximum trust.
-_ADMISSION_VERACITY_WEIGHT_PCT = {"approved_manual": 100, "approved_auto_safe": 90}
-_ADMISSION_VERACITY_DEFAULT_PCT = 90
 # Temporal query intent, conservative by construction: a fixed English cue
 # set (no per-language tables, per the routing-language policy) that only
 # doubles the RECENCY weight inside rank fusion. Relevance stays primary,
@@ -224,18 +298,13 @@ _ADMISSION_VERACITY_DEFAULT_PCT = 90
 # "latest", "now", and "newest" are ordinary engineering adjectives ("the
 # current implementation") and live in the phrase list instead, where the
 # surrounding words disambiguate them.
-_TEMPORAL_QUERY_CUES = frozenset({"yesterday", "today", "recent", "recently", "ago"})
-_TEMPORAL_QUERY_PHRASES = ("most recent", "right now", "as of now", "last week", "last month", "up to date")
 # Age tiers degrade the fused score of old records inside an equal relevance
 # rank, mnemosyne-style: 0-30 days full weight, 30-180 days half, older a
 # quarter. Relevance stays the primary key, so a stale strong match still
 # beats a fresh weak one; the tier only reorders peers.
-_AGE_TIER_BOUNDS_DAYS = (30, 180)
-_AGE_TIER_WEIGHTS = (1.0, 0.5, 0.25)
 # Pins are guaranteed-inclusion anchors, not eligibility overrides: a pinned
 # record still fails closed on expiry, scope, perspective, and review checks.
 # The cap stays small because pins occupy recall budget first.
-_MEMORY_PINS_LIMIT = 12
 # Attention tiers are Letta's context hierarchy read deterministically. A tier
 # says how much of the working context a record may occupy; it never says
 # whether the record is true, approved, or fresh. `active` is the working set,
@@ -246,9 +315,6 @@ _MEMORY_PINS_LIMIT = 12
 # `records/` and writes a tombstone; a tier change moves nothing and deletes
 # nothing. The tier feeds the one existing ranking ladder in
 # `build_project_memory_recall_pack`, never a second ordering pass.
-MEMORY_ATTENTION_TIERS = ("active", "reference", "archive")
-DEFAULT_MEMORY_ATTENTION_TIER = "active"
-_MEMORY_ATTENTION_RANK = {"active": 0, "reference": 1, "archive": 2}
 _RECALL_ATTENTION_KEYS = {
     "active_included",
     "reference_included",
@@ -303,7 +369,6 @@ _MEMORY_CONFIRMATION_CLAIM_BOUNDARY = (
 # record surfaces only through a matching lens, so an executor-specific
 # lesson never leaks into another executor's handoff.
 _PERSPECTIVE_KEYS = {"observer", "observed"}
-_DEFAULT_PERSPECTIVE_OBSERVER = "hermes"
 # Reciprocal rank fusion over deterministic signals, borrowed from hybrid
 # retrieval systems: heterogeneous signals are combined by rank, not by raw
 # score, so no signal needs scale normalization. Relevance rank stays the
@@ -311,8 +376,6 @@ _DEFAULT_PERSPECTIVE_OBSERVER = "hermes"
 # relevance rank, and it is stored as integer micro-units to stay valid
 # scalar metadata. Usage ranks on saturating buckets so delivery counts
 # cannot compound into a permanent head start.
-_RECALL_RRF_K = 60
-_RECALL_RRF_WEIGHTS = {"relevance": 2.0, "recency": 1.0, "usage": 1.0}
 _RECALL_USAGE_MAX_ENTRIES = 500
 _DERIVED_FROM_LIMIT = 8
 _LINEAGE_MAX_DEPTH = 10
@@ -340,58 +403,24 @@ _PROJECT_MEMORY_TASK_REF_KEYS = {"sha256", "length", "query_supplied"}
 # or unreadable file, or one past the cheap-hash budget -- reads as `unknown`
 # and never as `fresh`: a trust signal must fail closed, and "we could not
 # look" is not "we looked and it was fine".
-_SOURCE_EVIDENCE_MAX_BYTES = 4 * 1024 * 1024
 # Freshness warnings are the pre-handoff surface: a recall pack used to drop a
 # stale record silently, so the executor saw a smaller pack and no reason. The
 # list is bounded like every other polled surface; the pack already says
 # `truncated` when its own budget cuts records.
-_FRESHNESS_WARNING_LIMIT = 12
-_FRESHNESS_NEXT_ACTION = (
-    "Confirm, replace, or retire this record before it steers the plan; "
-    "`omh memory confirm <record-id>` resets its review deadline."
-)
 # The pre-deadline window turns the 90-day cliff into a slope: for this many
 # days before a record's review deadline, packs still deliver it but carry a
 # named warning, so the operator hears about the deadline while confirming
 # still keeps recall intact -- not on the first day the record is gone.
-_REVIEW_DUE_SOON_DAYS = 14
-_DUE_SOON_NEXT_ACTION = (
-    "Run `omh memory confirm <record-id>` (or correct/retire it) before the deadline passes; "
-    "until then the record still recalls normally."
-)
 # The retention twin of the review window: an expiring record (an episode's
 # 30-day TTL, most often) gets the same advance notice. Confirmation cannot
 # extend a TTL -- the honest actions are a correction with a longer TTL or
 # letting it expire and retiring it.
-_EXPIRES_SOON_DAYS = 14
-_EXPIRES_SOON_NEXT_ACTION = (
-    "Its retention TTL is about to end and confirmation cannot extend a TTL: "
-    "re-capture it (`omh memory capture --ttl-days N ...`) or correct it to keep the content, or let it expire."
-)
-_ADVISORY_NEXT_ACTIONS = {
-    "review_due_soon": _DUE_SOON_NEXT_ACTION,
-    "expires_soon": _EXPIRES_SOON_NEXT_ACTION,
-}
-_FRESHNESS_REASON_TEXT = {
-    "review_due_soon": "Its revalidation deadline is approaching; unconfirmed, it will leave default recall packs then.",
-    "expires_soon": "Its retention TTL is approaching; once it expires the record leaves recall entirely.",
-    "stale_review_required": "Its revalidation deadline passed, so nobody has confirmed the record since then.",
-    "source_changed": "The local source it cites changed after the record was approved.",
-    "source_unverifiable": "The local source it cites cannot be read now, so its freshness is unobservable.",
-    "superseded": "A newer revision supersedes this record.",
-    "expired_standard": "Its retention deadline passed.",
-    "expired_volatile": "Its retention deadline passed.",
-    "expired_durable": "Its retention deadline passed.",
-    "freshness_unconfirmed": "Its freshness could not be confirmed from stored metadata and local source evidence.",
-}
 # Reasons `--include-stale` may surface for inspection. They carry ineligible
 # replay evidence, so the pack still cannot be attached as approved context.
-_INSPECTABLE_STALE_REASONS = {"stale_review_required", "source_changed", "source_unverifiable"}
 # Advisory freshness reasons describe a record that is still fresh and still
 # delivered. Consumers that treat "has a freshness reason" as "this record is
 # stale" (wrapper continuity, role-context-pack diffs) must subtract this set,
 # or the advance notice would read as the failure it exists to prevent.
-ADVISORY_FRESHNESS_REASONS = frozenset({"review_due_soon", "expires_soon"})
 _HANDOFF_CONTEXT_PACK_KEYS = {
     "schema_version",
     "executor_target",
@@ -459,16 +488,6 @@ def build_project_memory_policy(paths: OmhPaths, *, mode: str | None = None) -> 
 # A notice window is not a retention period: a year is already generous, and
 # accepting the 100-year retention ceiling here would let one config typo turn
 # every record with any TTL into a permanent warning.
-_MEMORY_CADENCE_MAX = {"due_soon_days": 365}
-
-
-def _cadence_value(policy: dict[str, object], key: str) -> int | None:
-    """One validated cadence day-count from a policy mapping, or None."""
-    value = policy.get(key)
-    maximum = _MEMORY_CADENCE_MAX.get(key, MAX_RETENTION_DAYS)
-    if isinstance(value, int) and not isinstance(value, bool) and 1 <= value <= maximum:
-        return value
-    return None
 
 
 def _policy_cadence_overrides(stored_policy: dict[str, object]) -> dict[str, int]:
@@ -1221,34 +1240,6 @@ def reject_project_memory_candidate(
     }
 
 
-def effective_recall_configuration() -> dict[str, object]:
-    """Every constant the recall ladder ranks, decays, and cuts with.
-
-    A retrieval regression report is only comparable against another report
-    built under the same retrieval configuration, so the configuration has to
-    be readable rather than inferred from the sorted output. Reading the live
-    constants here -- not a copy of their values -- is the point: retuning a
-    weight changes this projection, which changes the report's configuration
-    digest, which is what makes two reports refuse to be compared.
-    """
-    return {
-        "recall_pack_schema_version": PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION,
-        "rrf_k": _RECALL_RRF_K,
-        "rrf_weights": dict(sorted(_RECALL_RRF_WEIGHTS.items())),
-        "temporal_recency_weight": 2.0,
-        "age_tier_bounds_days": list(_AGE_TIER_BOUNDS_DAYS),
-        "age_tier_weights": list(_AGE_TIER_WEIGHTS),
-        "attention_rank": dict(sorted(_MEMORY_ATTENTION_RANK.items())),
-        "admission_veracity_weight_pct": dict(sorted(_ADMISSION_VERACITY_WEIGHT_PCT.items())),
-        "admission_veracity_default_pct": _ADMISSION_VERACITY_DEFAULT_PCT,
-        "pins_limit": _MEMORY_PINS_LIMIT,
-        "inspectable_stale_reasons": sorted(_INSPECTABLE_STALE_REASONS),
-        "expires_soon_days": _EXPIRES_SOON_DAYS,
-        "review_due_soon_days": _REVIEW_DUE_SOON_DAYS,
-        "cadence_defaults": dict(sorted(_MEMORY_CADENCE_DEFAULTS.items())),
-    }
-
-
 def build_project_memory_recall_pack(
     paths: OmhPaths,
     query: str = "",
@@ -1268,204 +1259,61 @@ def build_project_memory_recall_pack(
     observer: str | None = None,
     observed: str | None = None,
     query_intent: str | None = None,
+    allowed_scopes: list[dict[str, str]] | tuple[dict[str, str], ...] | None = None,
+    required_scope_kinds: tuple[str, ...] | None = None,
+    inspection: bool = True,
 ) -> dict[str, object]:
-    policy = read_project_memory_policy(paths)
-    structural_selectors = (scope_kind, scope_ref, observer, observed)
-    if any(
-        value is not None and contains_credential_like_material(str(value))
-        for value in structural_selectors
-    ):
+    """Read the local store, then delegate every selection decision to the plugin.
+
+    Unscoped calls retain operator inspection semantics. Delivery callers use
+    explicit allowlists and inspection=False; partial selectors fail closed.
+    """
+    if any(value is not None and contains_credential_like_material(str(value)) for value in (scope_kind, scope_ref)):
         raise ValueError("credential-like recall selector is not allowed")
-    executor_target = _redacted_metadata_label(executor_target)
-    session_id = _redacted_metadata_label(session_id)
-    scope_kind = _redacted_metadata_label(scope_kind) if scope_kind is not None else None
-    scope_ref = _redacted_metadata_label(scope_ref) if scope_ref is not None else None
-    # Lens labels normalize exactly like capture labels: capture lowercases
-    # and strips, so a raw "--observed Codex" here would silently match
-    # nothing and read as "never captured".
-    observer = _redacted_metadata_label(str(observer or "").strip().lower()) or None
-    observed = _redacted_metadata_label(str(observed or "").strip().lower()) or None
-    task_ref = {
-        "sha256": hashlib.sha256(query.encode("utf-8")).hexdigest() if query else "",
-        "length": len(query),
-        "query_supplied": bool(query),
-    }
-    if not bool(policy.get("recall_enabled", True)):
-        return _empty_recall_pack(
-            policy,
-            executor_target=executor_target,
-            session_id=session_id,
-            task_ref=task_ref,
-            scope_kind=scope_kind,
-            scope_ref=scope_ref,
-            reason="project_memory_disabled",
-        )
+    if allowed_scopes is not None and (scope_kind is not None or scope_ref is not None):
+        raise ValueError("use either allowed_scopes or scope_kind/scope_ref")
+    if scope_kind is not None or scope_ref is not None:
+        allowed_scopes = [{"kind": str(scope_kind or ""), "ref": str(scope_ref or "")}]
+    required = required_scope_kinds if required_scope_kinds is not None else ((scope_kind,) if scope_kind else ("project",))
     records = _read_project_memory_records(paths)
-    requested_scope = _scope(scope_kind, scope_ref) if scope_kind and scope_ref else None
-    pins = set(read_memory_pins(paths))
-    review_resolver = _project_memory_review_resolver(paths)
-    included: list[dict[str, object]] = []
-    excluded: list[dict[str, object]] = []
-    archived_excluded = 0
+    return select_memory_recall(
+        records, query,
+        allowed_scopes=allowed_scopes,
+        required_scope_kinds=required,
+        inspection=inspection,
+        review_resolver=_project_memory_review_resolver(paths),
+        operation_states=_recall_operation_states(paths, records),
+        policy=read_project_memory_policy(paths),
+        usage=read_recall_usage(paths),
+        pins=set(read_memory_pins(paths)),
+        executor_target=executor_target,
+        session_id=session_id,
+        limit=limit,
+        max_chars=max_chars,
+        include_stale=include_stale,
+        include_archived=include_archived,
+        attention_override=attention_override,
+        now=now,
+        stale_override=stale_override,
+        run_id=run_id,
+        observer=observer,
+        observed=observed,
+        query_intent=query_intent,
+    ).pack
+
+
+def _recall_operation_states(paths: OmhPaths, records: list[dict[str, Any]]) -> dict[str, str]:
+    states: dict[str, str] = {}
     for record in records:
-        if not _record_scope_matches(record, scope_kind=scope_kind, scope_ref=scope_ref):
+        operation_id = record.get("operation_id")
+        if not isinstance(operation_id, str) or not operation_id or operation_id in states:
             continue
-        # Perspective mismatch skips silently, exactly like scope mismatch:
-        # the record belongs to another actor's lens, not to this pack.
-        if not _record_perspective_matches(record, observer=observer, observed=observed):
+        if not _SAFE_REF.fullmatch(operation_id):
+            states[operation_id] = "invalid"
             continue
-        attention_tier = _record_attention_tier(record, override=attention_override)
-        if attention_tier == "archive" and not include_archived:
-            # Archive is an attention tier, not a deletion. The record stays
-            # in `records/`, stays readable, and returns to any pack built
-            # with include_archived -- but it leaves the default working
-            # context NAMED, never silently, so the operator can always see
-            # what the archive is holding back.
-            excluded.append(
-                {
-                    "record_id": _redacted_metadata_label(record.get("record_id", "")),
-                    "reason": "archived_tier",
-                    "staleness": {"state": "not_checked"},
-                }
-            )
-            archived_excluded += 1
-            continue
-        evaluation = _evaluate_memory_artifact(
-            record,
-            paths=paths,
-            now=now,
-            requested_scope=requested_scope,
-            review_resolver=review_resolver,
-            stale_override=stale_override,
-            run_id=run_id,
-        )
-        staleness = _record_staleness(record, now=now, due_soon_days=_cadence_value(policy, "due_soon_days"))
-        # Source-evidence freshness gates recall exactly like the time
-        # deadline does. The shared evaluator only knows deadlines, so the
-        # verdict `_record_staleness` derived from the record's own recorded
-        # digest is folded into the same eligibility decision here rather
-        # than becoming a second, quieter notion of stale.
-        source_state = str(staleness.get("source_state", ""))
-        if bool(evaluation["eligible"]) and source_state in {"changed", "unreadable"}:
-            evaluation = {
-                **evaluation,
-                "eligible": False,
-                "reason_code": "source_changed" if source_state == "changed" else "source_unverifiable",
-            }
-        if not bool(evaluation["eligible"]):
-            # --include-stale is an inspection affordance: it surfaces
-            # records whose ONLY problem is unconfirmed freshness -- a passed
-            # revalidation deadline or a moved source -- carrying their
-            # ineligible replay evidence so the pack cannot be attached to a
-            # handoff. Expired and otherwise-ineligible records stay excluded
-            # regardless.
-            if include_stale and str(evaluation.get("reason_code", "")) in _INSPECTABLE_STALE_REASONS:
-                score = _memory_recall_score(record, query)
-                if not query or score > 0 or str(record.get("record_id", "")) in pins:
-                    included.append(
-                        _recall_item(record, score=score, staleness=staleness, evaluation=evaluation, attention_tier=attention_tier)
-                    )
-                    continue
-            excluded.append(_recall_exclusion(record, evaluation, staleness=staleness))
-            continue
-        score = _memory_recall_score(record, query)
-        # A pinned anchor is always in context: it skips only the
-        # no_query_overlap cut, never an eligibility check above.
-        if query and score <= 0 and str(record.get("record_id", "")) not in pins:
-            excluded.append(_recall_exclusion(record, evaluation, staleness=staleness, reason="no_query_overlap"))
-            continue
-        included.append(_recall_item(record, score=score, staleness=staleness, evaluation=evaluation, attention_tier=attention_tier))
-    query_intent = _resolve_query_intent(query, query_intent)
-    _attach_recall_ranking(
-        included,
-        read_recall_usage(paths),
-        pins=pins,
-        now=now if now is not None else datetime.now(timezone.utc),
-        recency_weight=2.0 if query_intent == "temporal" else _RECALL_RRF_WEIGHTS["recency"],
-    )
-    # Pinned anchors lead, then attention tier, then relevance; the decayed
-    # fused score orders records only within an equal relevance rank. A weaker
-    # keyword match can therefore never displace a stronger unpinned one of
-    # the same tier -- including across the budget cut below -- while recency,
-    # delivery usage, and age tier decide ties and unqueried packs. Attention
-    # sits above relevance on purpose: that is what "build recall packs from
-    # active records first" means, and it is the one place the tier acts.
-    # Pins take priority within the budget but never own it outright: at most
-    # limit-1 pinned slots lead the pack (minimum one), and further pins
-    # compete as normal records, so a fully-used pin budget cannot blank
-    # query-driven recall.
-    ranked_key = lambda item: (  # noqa: E731 - shared by both sort passes below
-        int(_ranking_field(item, "attention_rank")),
-        int(_ranking_field(item, "relevance_rank")),
-        -int(_ranking_field(item, "decayed_score_micro")),
-        str(item.get("record_id", "")),
-    )
-    privileged_pins = {
-        str(item.get("record_id", ""))
-        for item in sorted(
-            (item for item in included if _ranking_flag(item, "pinned")),
-            key=ranked_key,
-        )[: max(max(limit, 0) - 1, 1)]
-    }
-    included.sort(key=lambda item: (0 if str(item.get("record_id", "")) in privileged_pins else 1, *ranked_key(item)))
-    # Budget cut follows the priority ladder above: once either budget is
-    # crossed, everything after that point is cut, so a lower-priority record
-    # never displaces a higher-priority one. Cut records are recorded as
-    # over_budget rather than dropped silently -- the pack must be able to say
-    # "this is not everything".
-    kept: list[dict[str, object]] = []
-    kept_chars = 0
-    budget_exhausted = False
-    for item in included:
-        summary_chars = len(str(item.get("summary", "")))
-        if not budget_exhausted:
-            over_records = len(kept) >= max(limit, 0)
-            over_chars = max_chars is not None and kept_chars + summary_chars > max_chars
-            budget_exhausted = over_records or over_chars
-        if budget_exhausted:
-            entry = {
-                "record_id": str(item.get("record_id", "")),
-                "reason": "over_budget",
-                "staleness": item.get("staleness", {"state": "not_checked"}),
-                **_recall_evidence_fields(item.get("replay_evaluation")),
-            }
-            # A cut record that shares a tag with a KEPT record may be the
-            # other side of a same-topic disagreement; without this hint the
-            # surviving record silently "wins" until curation runs. The hint
-            # names the sibling only -- it never re-adds the record past the
-            # budget and never guesses which side is right.
-            cut_tags = {str(tag) for tag in item.get("tags", []) or []}
-            for kept_item in kept:
-                if cut_tags & {str(tag) for tag in kept_item.get("tags", []) or []}:
-                    entry["sibling_included"] = str(kept_item.get("record_id", ""))
-                    break
-            excluded.append(entry)
-            continue
-        kept.append(item)
-        kept_chars += summary_chars
-    included = kept
-    return {
-        "schema_version": PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION,
-        "enabled": True,
-        "executor_target": executor_target,
-        "session_id": session_id,
-        "task_ref": task_ref,
-        "policy": policy,
-        "scope": _scope(scope_kind or "project", scope_ref or "default"),
-        "perspective": {"observer": observer or "", "observed": observed or ""},
-        "query_intent": query_intent,
-        "included_records": included,
-        "excluded_records": excluded,
-        "freshness_warnings": _freshness_warnings(included, excluded),
-        "attention": _attention_disclosure(included, archived_excluded, include_archived=include_archived),
-        "record_count": len(included),
-        "truncated": budget_exhausted,
-        "redaction_policy": "metadata_only",
-        "claim_boundary": (
-            "Memory recall packs contain reviewed OMH project summaries only; "
-            "they are prepared context, not execution, review, CI, merge, or Hermes internal-memory evidence."
-        ),
-    }
+        operation, error = read_json_object_result(paths.memory_operations_dir / f"{operation_id}.json")
+        states[operation_id] = str(operation.get("state", "")) if not error and isinstance(operation, dict) else "unavailable"
+    return states
 
 
 def memory_recall_pack_for_handoff(
@@ -1481,13 +1329,17 @@ def memory_recall_pack_for_handoff(
     # pass as always, and records observed for this executor join them --
     # while a record about any other executor stays out of this pack.
     pack_scope = _handoff_pack_scope(paths, scope_kind=None, scope_ref=None)
+    allowed_scopes = [{"kind": "user-global", "ref": "default"}, pack_scope]
+    if session_id:
+        allowed_scopes.append({"kind": "thread", "ref": session_id})
     pack = build_project_memory_recall_pack(
         paths,
         query,
         executor_target=executor_target,
         session_id=session_id,
-        scope_kind=pack_scope["kind"],
-        scope_ref=pack_scope["ref"],
+        allowed_scopes=allowed_scopes,
+        required_scope_kinds=("project",),
+        inspection=False,
         limit=limit,
         observed=_handoff_perspective_lens(executor_target),
         # Per the routing-language policy the cue table stays English-only;
@@ -1608,113 +1460,6 @@ def record_recall_usage(paths: OmhPaths, record_ids: list[str], *, now: str | No
         "recorded": len(delivered),
         "records": {record_id: usage[record_id] for record_id in delivered},
     }
-
-
-def _ranking_field(item: dict[str, object], key: str) -> int:
-    ranking = item.get("ranking")
-    value = ranking.get(key, 0) if isinstance(ranking, dict) else 0
-    return value if isinstance(value, int) and not isinstance(value, bool) else 0
-
-
-def _ranking_flag(item: dict[str, object], key: str) -> bool:
-    ranking = item.get("ranking")
-    return bool(ranking.get(key, False)) if isinstance(ranking, dict) else False
-
-
-def _competition_ranks(items: list[dict[str, object]], value_fn: Any) -> dict[str, int]:
-    """1-based competition ranks, best first; equal values share a rank."""
-    ordered = sorted(items, key=lambda item: str(item.get("record_id", "")))
-    ordered.sort(key=value_fn, reverse=True)
-    ranks: dict[str, int] = {}
-    previous_value: object = object()
-    previous_rank = 1
-    for position, item in enumerate(ordered, start=1):
-        value = value_fn(item)
-        if value != previous_value:
-            previous_rank = position
-            previous_value = value
-        ranks[str(item.get("record_id", ""))] = previous_rank
-    return ranks
-
-
-def _attach_recall_ranking(
-    items: list[dict[str, object]],
-    usage: dict[str, dict[str, object]],
-    *,
-    pins: set[str] | None = None,
-    now: datetime | None = None,
-    recency_weight: float | None = None,
-) -> None:
-    """Fuse relevance, recency, and delivery usage into one recall order.
-
-    Without a query every relevance score ties at 1, so recency and usage
-    decide the order instead of the record-id accident the pure keyword sort
-    fell back to. Recency ranks on the approved_at ISO string, which sorts
-    lexicographically; a record missing approved_at ranks oldest. The fused
-    score is then degraded by age tier so old records reorder below young
-    peers of equal relevance.
-    """
-    if not items:
-        return
-    pins = pins or set()
-    now = now if now is not None else datetime.now(timezone.utc)
-    recency_weight = recency_weight if recency_weight is not None else _RECALL_RRF_WEIGHTS["recency"]
-    def _times_recalled(item: dict[str, object]) -> int:
-        entry = usage.get(str(item.get("record_id", "")), {})
-        times = entry.get("times_recalled", 0)
-        return times if isinstance(times, int) and not isinstance(times, bool) else 0
-
-    relevance = _competition_ranks(items, lambda item: int(item.get("score", 0) or 0))
-    recency = _competition_ranks(items, lambda item: str(item.get("approved_at", "")))
-    usage_ranks = _competition_ranks(items, lambda item: _usage_bucket(_times_recalled(item)))
-    for item in items:
-        record_id = str(item.get("record_id", ""))
-        fused = (
-            _RECALL_RRF_WEIGHTS["relevance"] / (_RECALL_RRF_K + relevance[record_id])
-            + recency_weight / (_RECALL_RRF_K + recency[record_id])
-            + _RECALL_RRF_WEIGHTS["usage"] / (_RECALL_RRF_K + usage_ranks[record_id])
-        )
-        tier = _age_tier(str(item.get("approved_at", "")), now=now)
-        veracity_pct = _ADMISSION_VERACITY_WEIGHT_PCT.get(str(item.get("admission_mode", "")), _ADMISSION_VERACITY_DEFAULT_PCT)
-        # The attention rank is an ordering key, not a score input: it never
-        # touches the fused score, so a tier change reorders records without
-        # rewriting the relevance/recency/usage evidence that explains them.
-        attention_rank = _MEMORY_ATTENTION_RANK.get(
-            str(item.get("attention_tier", "")) or DEFAULT_MEMORY_ATTENTION_TIER,
-            _MEMORY_ATTENTION_RANK[DEFAULT_MEMORY_ATTENTION_TIER],
-        )
-        item["ranking"] = {
-            # rrf_score_micro is undecayed rank fusion under THIS pack's
-            # weights (a temporal query raises the recency weight, so scores
-            # compare within a pack, not across packs); age decay and
-            # veracity land in decayed_score_micro, which is what the sort
-            # uses.
-            "rrf_score_micro": round(fused * 1_000_000),
-            "decayed_score_micro": round(fused * _AGE_TIER_WEIGHTS[tier] * (veracity_pct / 100) * 1_000_000),
-            "relevance_rank": relevance[record_id],
-            "recency_rank": recency[record_id],
-            "usage_rank": usage_ranks[record_id],
-            "times_recalled": _times_recalled(item),
-            "age_tier": tier,
-            "attention_rank": attention_rank,
-            "pinned": record_id in pins,
-            "veracity_weight_pct": veracity_pct,
-        }
-
-
-def _usage_bucket(times_recalled: int) -> int:
-    """Saturating ordinal for the usage signal: 0, 1-2, 3-9, 10+.
-
-    Raw counts self-reinforce -- every delivery improves the rank that earns
-    the next delivery -- so the signal saturates instead of compounding.
-    """
-    if times_recalled >= 10:
-        return 3
-    if times_recalled >= 3:
-        return 2
-    if times_recalled >= 1:
-        return 1
-    return 0
 
 
 def _normalized_summary_key(summary: str) -> str:
@@ -1922,53 +1667,6 @@ def _find_pending_candidate_by_summary(paths: OmhPaths, summary: str) -> str:
     return ""
 
 
-def _recall_query_intent(query: str) -> str:
-    """Return "temporal" for an unambiguous time cue, else "default".
-
-    Token cues use the recall tokenizer, so matching is exact token overlap,
-    never substring guessing ("nowhere" and "knownHosts" stay default).
-    Known limitation: hyphenated compounds tokenize whole ("recently-requested"),
-    so a cue inside one does not fire. Phrase cues match on whitespace-
-    normalized lowercase text.
-    """
-    if not query.strip():
-        return "default"
-    if _memory_tokens(query) & _TEMPORAL_QUERY_CUES:
-        return "temporal"
-    normalized = f" {' '.join(query.lower().split())} "
-    return "temporal" if any(f" {phrase} " in normalized for phrase in _TEMPORAL_QUERY_PHRASES) else "default"
-
-
-def _resolve_query_intent(query: str, supplied: str | None) -> str:
-    """The caller's stated intent when there is one, else the English cues.
-
-    The cue table is English-only and stays that way. Per the routing-language
-    policy (`tests/test_routing_language_policy.py`, `src/routing/input_language.py`)
-    per-language trigger tables do not scale to a global product, and non-English
-    intent resolution belongs to model selection over supplied candidates rather
-    than to more tokens. Measured before this existed: every Korean and Japanese
-    phrasing of "recently" -- `어제 뭐 정했지`, `최근 배포 결정`, `3일 전에 정한 거`,
-    `最近の変更` -- resolved to `default`, so recency weighting never engaged for
-    them while `recent changes` got it.
-
-    Adding those words to the table would have fixed five languages and left the
-    rest, which is the habit the policy exists to stop. So the caller states it
-    instead: Hermes read the message and already knows whether the user asked
-    for the latest, in any language, and now has somewhere to say so.
-
-    An unrecognized value is refused rather than ignored. A caller that
-    misspells its intent should learn that, not silently get the default.
-    """
-    if supplied is None:
-        return _recall_query_intent(query)
-    normalized = str(supplied).strip().lower()
-    if normalized in {"", "auto"}:
-        return _recall_query_intent(query)
-    if normalized not in {"default", "temporal"}:
-        raise ValueError("query_intent must be one of auto, default, temporal")
-    return normalized
-
-
 def _memory_pins_path(paths: OmhPaths) -> Path:
     return paths.memory_dir / "pins.json"
 
@@ -2026,45 +1724,6 @@ def set_memory_pin(paths: OmhPaths, record_id: str, *, pinned: bool) -> dict[str
     }
 
 
-def normalize_memory_attention_tier(value: Any) -> str:
-    """The one place a tier name is accepted. An unknown tier is refused.
-
-    Failing loudly here is deliberate: silently coercing an unrecognized tier
-    to ``active`` would quietly undo an operator's archive request.
-    """
-    tier = str(value or "").strip().lower()
-    if tier not in MEMORY_ATTENTION_TIERS:
-        raise ValueError(f"unsupported memory attention tier: {value!r}; use one of {', '.join(MEMORY_ATTENTION_TIERS)}")
-    return tier
-
-
-def record_attention_tier(record: dict[str, Any]) -> str:
-    """A stored record's attention tier; anything unreadable reads as active.
-
-    Absence is the normal case for every record approved before tiers existed,
-    and a corrupt tier value is indistinguishable from absence here. Defaulting
-    to ``active`` is safe because attention is not a trust gate: expiry, scope,
-    perspective, and review eligibility still decide what may be recalled.
-    """
-    attention = record.get("attention")
-    tier = str(attention.get("tier", "")) if isinstance(attention, dict) else ""
-    return tier if tier in MEMORY_ATTENTION_TIERS else DEFAULT_MEMORY_ATTENTION_TIER
-
-
-def _record_attention_tier(record: dict[str, Any], *, override: dict[str, str] | None = None) -> str:
-    """Stored tier, or the previewed tier when this pack is a projection.
-
-    The override exists so a preview and the pack built after apply run the
-    exact same ranking code on the exact same inputs. Recomputing the order a
-    second way would make "what will remain in the working context" a guess.
-    """
-    if override:
-        candidate = override.get(str(record.get("record_id", "")))
-        if candidate is not None:
-            return normalize_memory_attention_tier(candidate)
-    return record_attention_tier(record)
-
-
 def _attention_metadata(tier: str, *, reason: str, previous_tier: str, changed_at: str) -> dict[str, str]:
     """Scalar-only tier metadata; nested non-scalars would be dropped by
     correction/restore and by the recall-pack validators."""
@@ -2074,47 +1733,6 @@ def _attention_metadata(tier: str, *, reason: str, previous_tier: str, changed_a
         "reason": _redact(str(reason or ""))[:_MEMORY_ATTENTION_REASON_LIMIT],
         "previous_tier": str(previous_tier or ""),
         "changed_at": str(changed_at or ""),
-    }
-
-
-def _attention_disclosure(
-    included: list[dict[str, object]],
-    archived_excluded: int,
-    *,
-    include_archived: bool,
-) -> dict[str, object]:
-    """Say out loud which tiers this pack is made of.
-
-    The issue asks recall to disclose when reference records are included, and
-    the same sentence is the only honest way to report an archived record that
-    was held back: a smaller pack with no explanation is the failure this
-    replaces.
-    """
-    counts = dict.fromkeys(MEMORY_ATTENTION_TIERS, 0)
-    for item in included:
-        tier = str(item.get("attention_tier", "")) or DEFAULT_MEMORY_ATTENTION_TIER
-        counts[tier if tier in counts else DEFAULT_MEMORY_ATTENTION_TIER] += 1
-    parts: list[str] = []
-    if not included:
-        parts.append("No reviewed records are in the working context.")
-    else:
-        parts.append(f"{counts['active']} active record(s) lead this working context.")
-    if counts["reference"]:
-        parts.append(f"{counts['reference']} reference-tier record(s) are included behind them.")
-    if counts["archive"]:
-        parts.append(f"{counts['archive']} archived record(s) are included because this query asked for the archive.")
-    if archived_excluded:
-        parts.append(
-            f"{archived_excluded} archived record(s) stayed out of the working context; "
-            "they remain in the store and are listed as archived_tier exclusions."
-        )
-    return {
-        "active_included": counts["active"],
-        "reference_included": counts["reference"],
-        "archived_included": counts["archive"],
-        "archived_excluded": int(archived_excluded),
-        "include_archived": bool(include_archived),
-        "detail": " ".join(parts),
     }
 
 
@@ -2553,23 +2171,6 @@ def _refused_confirmation(record_id: str, reason_code: str) -> dict[str, object]
     }
 
 
-def _age_tier(approved_at: str, *, now: datetime) -> int:
-    """0 for young, 1 for aging, 2 for old; unparseable timestamps stay 0 so
-    a malformed record is never silently downweighted."""
-    # Naive timestamps read as UTC, matching the expiry classifier: the
-    # plain _parse_utc would read them as host-local and shift the tier by
-    # up to +/-14 hours at the 30/180-day boundaries.
-    approved = _parse_utc_naive_as_utc(str(approved_at or ""))
-    if approved is None:
-        return 0
-    age_days = max((now - approved).total_seconds(), 0.0) / 86400.0
-    if age_days >= _AGE_TIER_BOUNDS_DAYS[1]:
-        return 2
-    if age_days >= _AGE_TIER_BOUNDS_DAYS[0]:
-        return 1
-    return 0
-
-
 def _handoff_perspective_lens(executor_target: str | None) -> str:
     """Lens every handoff surface applies: scoped records reach only the
     executor they are about. An unresolved target ('' or 'choose') stays a
@@ -2612,39 +2213,6 @@ def _normalize_perspective(observer: str | None, observed: str | None) -> dict[s
         if not _SAFE_REF.match(label):
             raise ValueError(f"unsafe perspective actor label: {label!r}")
     return {"observer": observer_label, "observed": observed_label}
-
-
-def _perspective_projection(value: Any) -> dict[str, str]:
-    """Projection of a stored perspective; {} when there is nothing observed.
-
-    An empty observed actor is not a perspective -- the lens ignores it -- so
-    projecting it would advertise scoping the filter does not apply.
-    """
-    perspective = value if isinstance(value, dict) else {}
-    observed = str(perspective.get("observed", ""))
-    if not observed:
-        return {}
-    return {"observer": str(perspective.get("observer", "") or _DEFAULT_PERSPECTIVE_OBSERVER), "observed": observed}
-
-
-def _record_perspective_matches(record: dict[str, Any], *, observer: str | None, observed: str | None) -> bool:
-    """Lens semantics: unscoped records always pass; scoped need a match.
-
-    No lens (both None) also passes scoped records -- a plain recall is an
-    inspection surface and hides nothing. Filtering both ways (a lens that
-    excludes unscoped records) is deliberately not offered: unscoped is the
-    compatibility default, not a perspective of its own.
-    """
-    perspective = record.get("perspective")
-    if not isinstance(perspective, dict) or not str(perspective.get("observed", "")):
-        return True
-    if observer is None and observed is None:
-        return True
-    if observer is not None and str(perspective.get("observer", "")) != observer:
-        return False
-    if observed is not None and str(perspective.get("observed", "")) != observed:
-        return False
-    return True
 
 
 def build_memory_perspectives(paths: OmhPaths) -> dict[str, object]:
@@ -3702,174 +3270,6 @@ def _staleness_projection(revalidation: dict[str, object]) -> dict[str, object]:
     return {"stale_after": deadline, "stale_after_days": None, "review_due_at": deadline}
 
 
-def _empty_recall_pack(
-    policy: dict[str, object],
-    *,
-    executor_target: str,
-    session_id: str,
-    task_ref: dict[str, object],
-    scope_kind: str | None,
-    scope_ref: str | None,
-    reason: str,
-) -> dict[str, object]:
-    return {
-        "schema_version": PROJECT_MEMORY_RECALL_PACK_SCHEMA_VERSION,
-        "enabled": False,
-        "executor_target": executor_target,
-        "session_id": session_id,
-        "task_ref": task_ref,
-        "policy": policy,
-        "scope": _scope(scope_kind or "project", scope_ref or "default"),
-        "perspective": {"observer": "", "observed": ""},
-        "query_intent": "default",
-        "included_records": [],
-        "excluded_records": [{"record_id": "", "reason": reason, "staleness": {"state": "not_checked"}}],
-        "freshness_warnings": [],
-        "attention": _attention_disclosure([], 0, include_archived=False),
-        "record_count": 0,
-        "truncated": False,
-        "redaction_policy": "metadata_only",
-        "claim_boundary": "Memory recall is disabled or empty; no execution, review, CI, merge, or Hermes internal-memory evidence is produced.",
-    }
-
-
-def _recall_item(
-    record: dict[str, Any],
-    *,
-    score: int,
-    staleness: dict[str, object],
-    evaluation: dict[str, object],
-    attention_tier: str = DEFAULT_MEMORY_ATTENTION_TIER,
-) -> dict[str, object]:
-    evidence = _replay_evaluation(record, evaluation)
-    return {
-        "record_id": _redacted_metadata_label(record.get("record_id", "")),
-        "record_type": _redact_admitted_text(str(record.get("record_type", ""))),
-        "summary": _redact_admitted_text(str(record.get("summary", "")))[:500],
-        "scope": _normalize_scope(record.get("scope", _scope("project", "default"))),
-        "tags": _normalize_tags(record.get("tags", [])),
-        "source": str(record.get("source", "")),
-        "approved_at": _redact_admitted_text(str(record.get("approved_at", ""))),
-        "staleness": _redact_nested_metadata(staleness),
-        "score": int(score),
-        "attention_tier": attention_tier,
-        "derived_from": _string_list(record.get("derived_from", [])),
-        "perspective": _perspective_projection(record.get("perspective")),
-        **_recall_evidence_fields(evidence),
-    }
-
-
-def _recall_exclusion(
-    record: dict[str, Any],
-    evaluation: dict[str, object],
-    *,
-    staleness: dict[str, object],
-    reason: str | None = None,
-) -> dict[str, object]:
-    evidence = _replay_evaluation(record, evaluation)
-    return {
-        "record_id": _redacted_metadata_label(record.get("record_id", "")),
-        "reason": reason or str(evidence["reason_code"]),
-        "staleness": _redact_nested_metadata(staleness),
-        **_recall_evidence_fields(evidence),
-    }
-
-
-def _freshness_warnings(
-    included: list[dict[str, object]],
-    excluded: list[dict[str, object]],
-) -> list[dict[str, object]]:
-    """Name every record in this pack whose freshness is not confirmed.
-
-    This is the surface acceptance criterion 1 asks for: before a stale,
-    expired, superseded, or source-moved record can influence a plan or a
-    handoff, the pack says which record it is, why revalidation is due, and
-    that the operator must confirm, replace, or retire it. A held-back record
-    warns exactly like a delivered one -- ``delivered`` says which happened --
-    because a record silently missing from a pack is the failure this
-    replaces, not the fix.
-
-    Records excluded for reasons that are not about freshness
-    (``no_query_overlap``, ``over_budget`` on a fresh record) never warn:
-    a warning that fires for everything is read as noise and stops working.
-    """
-    blocking: list[dict[str, object]] = []
-    advisory: list[dict[str, object]] = []
-    seen: set[str] = set()
-    for delivered, entry in [*((True, item) for item in included), *((False, item) for item in excluded)]:
-        record_id = str(entry.get("record_id", ""))
-        staleness = entry.get("staleness") if isinstance(entry.get("staleness"), dict) else {}
-        state = str(staleness.get("state", ""))
-        reason_code = str(entry.get("eligibility_reason", "") or "")
-        known_reason = reason_code in _FRESHNESS_REASON_TEXT
-        staleness_reason = str(staleness.get("reason", "") or "")
-        if not known_reason and delivered and staleness_reason in ADVISORY_FRESHNESS_REASONS:
-            # A still-fresh record inside a pre-deadline window (review-due or
-            # TTL expiry): eligibility has nothing to say about it, so the
-            # advance notice comes from the staleness verdict instead. Only a
-            # DELIVERED record earns it -- an advisory on every unrelated
-            # record in the store would page the operator about records this
-            # task never touched, and once the deadline actually passes the
-            # ordinary blocking warning fires for held-back records as before.
-            reason_code, known_reason = staleness_reason, True
-        if not known_reason and state in {"", "fresh", "not_checked"}:
-            continue
-        if not record_id or record_id in seen:
-            continue
-        seen.add(record_id)
-        if not known_reason:
-            reason_code = "freshness_unconfirmed"
-        advisory_reason = reason_code in ADVISORY_FRESHNESS_REASONS
-        # The limit truncates advisory notices first: an advance heads-up must
-        # never displace the name of a record that actually left the pack --
-        # naming those is the guarantee this list exists to keep.
-        (advisory if advisory_reason else blocking).append(
-            {
-                "record_id": record_id,
-                "state": state or "unknown",
-                "reason_code": reason_code,
-                "review_due_at": str(staleness.get("review_due_at", "") or ""),
-                "expires_at": str(staleness.get("expires_at", "") or ""),
-                "detail": _FRESHNESS_REASON_TEXT[reason_code],
-                "delivered": bool(delivered),
-                "next_action": _ADVISORY_NEXT_ACTIONS.get(reason_code, _FRESHNESS_NEXT_ACTION),
-            }
-        )
-        if len(blocking) >= _FRESHNESS_WARNING_LIMIT:
-            break
-    return (blocking + advisory)[:_FRESHNESS_WARNING_LIMIT]
-
-
-def freshness_reason_detail(reason_code: str) -> str:
-    """Human text for one recall-pack freshness reason code, or "" when unknown.
-
-    Public so other handoff surfaces can explain a record with the vocabulary
-    the recall pack already emits instead of growing a parallel table that
-    drifts from it. Read-only on purpose: adding a code to `_FRESHNESS_REASON_TEXT`
-    also changes what `_freshness_warnings` treats as a freshness reason, so
-    the table stays owned by recall. Membership no longer implies stale:
-    codes in `ADVISORY_FRESHNESS_REASONS` describe a still-fresh record, and
-    consumers that map "has a reason" to "is stale" must subtract that set.
-    """
-    return _FRESHNESS_REASON_TEXT.get(str(reason_code), "")
-
-
-def _recall_evidence_fields(value: Any) -> dict[str, object]:
-    evidence = value if isinstance(value, dict) else {}
-    safe_evidence = _redact_nested_metadata(evidence)
-    evidence = safe_evidence if isinstance(safe_evidence, dict) else {}
-    return {
-        "revision": int(evidence.get("revision", 0) or 0),
-        "admission_mode": str(evidence.get("admission_mode") or ""),
-        "source_class": str(evidence.get("source_class") or ""),
-        "retention_class": str(evidence.get("retention_class") or ""),
-        "evaluated_at": str(evidence.get("evaluated_at") or ""),
-        "eligibility_reason": str(evidence.get("reason_code") or ""),
-        "revalidation_evidence": evidence.get("revalidation_evidence", {}),
-        "replay_evaluation": evidence,
-    }
-
-
 def _project_memory_review_resolver(paths: OmhPaths) -> dict[str, dict[str, object]]:
     return {
         str(review.get("review_id", "")): review
@@ -3889,14 +3289,9 @@ def _evaluate_memory_artifact(
     stale_override: dict[str, object] | None = None,
     run_id: str | None = None,
 ) -> dict[str, object]:
-    evaluator_artifact = _normalize_evaluator_timestamps(artifact)
-    if evaluator_artifact.get("schema_version") == LEGACY_MEMORY_SCOPE_SCHEMA_VERSION:
-        # Governance has one legacy reason code for v1 project records. Map
-        # the preserved v1 scope schema into that read-only compatibility
-        # classification rather than coercing it into an invalid v2 artifact.
-        evaluator_artifact = {**evaluator_artifact, "schema_version": LEGACY_PROJECT_MEMORY_RECORD_SCHEMA_VERSION}
-    result = evaluate_memory_replay(
-        evaluator_artifact,
+    return _shared_evaluate_memory_artifact(
+        artifact,
+        operation_states=_recall_operation_states(paths, [artifact]) if paths is not None else None,
         now=now,
         requested_scope=requested_scope,
         review_resolver=review_resolver,
@@ -3904,102 +3299,10 @@ def _evaluate_memory_artifact(
         stale_override=stale_override,
         run_id=run_id,
     )
-    admission = artifact.get("admission")
-    review_id = admission.get("review_id") if isinstance(admission, dict) else ""
-    if (
-        result.get("eligible") is True
-        and artifact.get("schema_version") in {PROJECT_MEMORY_RECORD_SCHEMA_VERSION, MEMORY_SCOPE_SCHEMA_VERSION}
-        and (not isinstance(review_id, str) or not review_id or not review_resolver or review_id not in review_resolver)
-    ):
-        # The core boundary supplies a resolver, so an approval without its
-        # immutable review record cannot become eligible through an omitted id.
-        result = {**result, "eligible": False, "reason_code": "review_not_found"}
-    operation_id = artifact.get("operation_id")
-    if result.get("eligible") is True and paths and isinstance(operation_id, str) and operation_id:
-        operation, error = read_json_object_result(paths.memory_operations_dir / f"{operation_id}.json")
-        if error or not isinstance(operation, dict) or operation.get("state") != "completed":
-            result = {**result, "eligible": False, "reason_code": "operation_incomplete"}
-    return _replay_evaluation(artifact, result)
 
 
-def _normalize_evaluator_timestamps(artifact: dict[str, Any]) -> dict[str, Any]:
-    """Present legacy-naive ISO deadlines to the shared evaluator as UTC."""
-    normalized = dict(artifact)
-    for field, timestamp_key in (("retention", "expires_at"), ("revalidation", "deadline")):
-        metadata = artifact.get(field)
-        if not isinstance(metadata, dict) or not metadata.get(timestamp_key):
-            continue
-        parsed = _parse_utc_naive_as_utc(str(metadata[timestamp_key]))
-        if parsed is None:
-            continue
-        normalized[field] = {**metadata, timestamp_key: parsed.isoformat().replace("+00:00", "Z")}
-    return normalized
-
-
-def _parse_utc_naive_as_utc(value: str) -> datetime | None:
-    try:
-        parsed = datetime.fromisoformat(value.replace("Z", "+00:00"))
-    except ValueError:
-        return None
-    if parsed.tzinfo is None:
-        parsed = parsed.replace(tzinfo=timezone.utc)
-    return parsed.astimezone(timezone.utc)
-
-
-def _replay_evaluation(artifact: dict[str, Any], result: dict[str, object]) -> dict[str, object]:
-    try:
-        identity = stable_artifact_identity(artifact)
-    except ValueError:
-        identity = {}
-    revalidation = artifact.get("revalidation")
-    return {
-        "schema_version": "omh_memory_replay_evaluation/v1",
-        "artifact_identity": identity,
-        "revision": int(artifact.get("revision", 0) or 0),
-        "admission_mode": str(result.get("admission_mode") or ""),
-        "source_class": str(result.get("source_class") or ""),
-        "retention_class": str(result.get("retention_class") or _retention_class(artifact)),
-        "evaluated_at": str(result.get("evaluated_at", "")),
-        "eligible": bool(result.get("eligible", False)),
-        "reason_code": str(result.get("reason_code", "unknown")),
-        "revalidation_evidence": {"deadline": str(revalidation.get("deadline", ""))} if isinstance(revalidation, dict) else {},
-    }
-
-
-def _retention_class(artifact: dict[str, Any]) -> str:
-    retention = artifact.get("retention")
-    return str(retention.get("class", "")) if isinstance(retention, dict) else ""
-
-
-def _memory_recall_score(record: dict[str, Any], query: str) -> int:
-    if not query.strip():
-        return 1
-    query_tokens = _memory_tokens(query)
-    if not query_tokens:
-        # The query carries no indexable tokens at all (emoji-only, an
-        # unsupported script, or only sub-length words). Scoring it as zero
-        # overlap used to exclude every record as no_query_overlap and hand
-        # the executor an empty pack; fall back to unqueried recall so the
-        # budget ladder still surfaces approved records.
-        return 1
-    record_tokens = _memory_tokens(
-        " ".join(
-            [
-                str(record.get("summary", "")),
-                str(record.get("record_type", "")),
-                " ".join(_normalize_tags(record.get("tags", []))),
-            ]
-        )
-    )
-    overlap = query_tokens & record_tokens
-    tag_overlap = query_tokens & set(_normalize_tags(record.get("tags", [])))
-    return len(overlap) * 10 + len(tag_overlap) * 5
-
-
-_MEMORY_ASCII_TOKEN = re.compile(r"[a-z0-9_/-]{3,}")
 # A whole two-character word, never a fragment of a longer one: the lookarounds
 # keep "ru" out of "runs" while letting "ci" out of "ci failures".
-_MEMORY_SHORT_ASCII_TOKEN = re.compile(r"(?<![a-z0-9_/-])[a-z0-9]{2}(?![a-z0-9_/-])")
 # The length floor was doing two jobs: keeping English function words out of the
 # index, and -- as a side effect nobody chose -- keeping every two-letter
 # technical term out with them. Only the first job is wanted, so it is done by
@@ -4009,150 +3312,11 @@ _MEMORY_SHORT_ASCII_TOKEN = re.compile(r"(?<![a-z0-9_/-])[a-z0-9]{2}(?![a-z0-9_/
 # `go`, `id`, `db`, `ci`, `ui`, `qa`, `ml`, `pr`, `js`, `ts`, `vm`, `s3`, `k8`,
 # `ai`, `ux`, and `vs` are all real terms in some project and stay indexable: a
 # stopword that costs a real search is worse than the noise it removes.
-_MEMORY_SHORT_STOPWORDS = frozenset(
-    {
-        "am", "an", "as", "at", "be", "by", "do", "he", "if", "in", "is", "it",
-        "me", "my", "no", "of", "oh", "on", "or", "so", "to", "up", "us", "we",
-    }
-)
-_MEMORY_CJK_RUN = re.compile(r"[\u3040-\u30ff\u3400-\u4dbf\u4e00-\u9fff\uac00-\ud7a3]+")
-
-
-def _memory_tokens(value: str) -> set[str]:
-    """Index tokens for recall scoring, covering ASCII and CJK text.
-
-    ASCII words of three characters or more are indexed whole. Two-character
-    words are indexed unless they are one of the named function words. The old
-    >=3 floor made every two-letter technical term unreachable, and the two
-    failure modes it produced were opposites: "CI" tokenized to nothing, so the
-    no-indexable-tokens fallback handed back the whole store with no exclusion
-    reason, while "ci failures" tokenized to {failures}, so every record came
-    back no_query_overlap -- including one whose summary began with "CI" and
-    carried `ci` as a tag, which the tag bonus could not rescue because the
-    query token had already been dropped.
-
-    CJK runs are indexed as the whole
-    run plus its character bigrams: Korean particles glue to the noun
-    ("배포는"), so whole-word overlap alone would miss "배포" in a query.
-    The previous ASCII-only split tokenized any CJK query to the empty set,
-    which excluded every record as no_query_overlap and silently emptied
-    recall packs for projects that chat in Korean, Japanese, or Chinese.
-    """
-    lowered = unicodedata.normalize("NFC", value).lower()
-    tokens = set(_MEMORY_ASCII_TOKEN.findall(lowered))
-    tokens.update(token for token in _MEMORY_SHORT_ASCII_TOKEN.findall(lowered) if token not in _MEMORY_SHORT_STOPWORDS)
-    for run in _MEMORY_CJK_RUN.findall(lowered):
-        if len(run) >= 2:
-            tokens.add(run)
-        tokens.update(run[index : index + 2] for index in range(len(run) - 1))
-    return tokens
 
 
 def _record_scope_matches(record: dict[str, Any], *, scope_kind: str | None, scope_ref: str | None) -> bool:
     scope = _normalize_scope(record.get("scope", _scope("project", "default")))
     return (not scope_kind or scope["kind"] == scope_kind) and (not scope_ref or scope["ref"] == scope_ref)
-
-
-def _record_staleness(
-    record: dict[str, Any],
-    *,
-    now: datetime | None = None,
-    due_soon_days: int | None = None,
-) -> dict[str, object]:
-    """The one freshness verdict: TTL, review-due date, and source evidence.
-
-    The TTL half is decided by the bundle classifier, the single source of
-    truth for what "expired" means: it reads naive timestamps as UTC, where
-    the local ``_parse_utc`` would read them as host-local time and move the
-    verdict by up to +/-14 hours depending on where the host happens to be.
-
-    The review-due half reads ``review_due_at`` and the older ``stale_after``
-    spelling of the same date, taking whichever comes first, so records
-    written before the field was named keep their deadline and a record with
-    only one spelling edited cannot read as fresh.
-
-    The source half compares the digest recorded at capture against the cited
-    local file as it reads now. It only ever makes a record less trusted: a
-    moved source is ``stale``, an unreadable one is ``unknown``, and neither
-    can turn into ``fresh``. Every input is stored metadata plus the caller's
-    ``now`` plus locally observable bytes, so the verdict is reproducible.
-    """
-    now = now if now is not None else datetime.now(timezone.utc)
-    ttl = record.get("ttl", {}) if isinstance(record.get("ttl"), dict) else {}
-    staleness = record.get("staleness", {}) if isinstance(record.get("staleness"), dict) else {}
-    expires_at = str(ttl.get("expires_at", ""))
-    stale_after = str(staleness.get("stale_after", ""))
-    review_due_at = _earliest_deadline(str(staleness.get("review_due_at", "") or ""), stale_after)
-    source_state = _source_evidence_state(record)
-    fields = {
-        "stale_after": stale_after,
-        "review_due_at": review_due_at,
-        "expires_at": expires_at,
-        "source_state": source_state,
-    }
-    if _classify_record_expiry(record, now=now) == "expired":
-        return {"state": "expired", "reason": "retention_expired", **fields}
-    deadline = _parse_utc(review_due_at)
-    if deadline and deadline <= now:
-        return {"state": "stale", "reason": "review_due", **fields}
-    if source_state == "changed":
-        return {"state": "stale", "reason": "source_changed", **fields}
-    if source_state == "unreadable":
-        return {"state": "unknown", "reason": "source_unreadable", **fields}
-    # Still fresh and still eligible below here: the record delivers exactly
-    # as before. The reason is the advance notice recall packs turn into a
-    # warning, so a deadline stops being a surprise discovered only after the
-    # record has already left every pack. Expiry outranks review-due when both
-    # windows overlap: a passed TTL is terminal, a passed review date is not.
-    # Naive TTL timestamps read as UTC, matching the expiry classifier that
-    # decides the terminal verdict above; `_parse_utc` would read them as
-    # host-local and move this window by up to +/-14 hours.
-    expiry = _parse_utc_naive_as_utc(expires_at)
-    expiry_window = due_soon_days if due_soon_days is not None else _EXPIRES_SOON_DAYS
-    # A notice window longer than half the record's own life is not advance
-    # notice, it is a permanent banner: a 7-day volatile record would warn
-    # from the moment it was approved. Scale the window down to half the
-    # record's lifespan (never below one day) so short-lived records warn
-    # near the end, which is when the warning means something. The span
-    # comes from ttl_days when there is one, else from the distance between
-    # creation and the absolute expiry date -- an absolute two-day deadline
-    # must not warn from birth either.
-    ttl_days_value = ttl.get("ttl_days")
-    span_days: int | None = None
-    if isinstance(ttl_days_value, int) and not isinstance(ttl_days_value, bool) and ttl_days_value > 0:
-        span_days = ttl_days_value
-    elif expiry is not None:
-        created = _parse_utc_naive_as_utc(str(record.get("created_at", "") or ""))
-        if created is not None and expiry > created:
-            span_days = max(int((expiry - created).total_seconds() // 86400), 1)
-    if span_days is not None:
-        expiry_window = min(expiry_window, max(span_days // 2, 1))
-    if expiry and expiry - now <= timedelta(days=expiry_window):
-        return {"state": "fresh", "reason": "expires_soon", **fields}
-    if deadline and deadline - now <= timedelta(days=due_soon_days if due_soon_days is not None else _REVIEW_DUE_SOON_DAYS):
-        return {"state": "fresh", "reason": "review_due_soon", **fields}
-    return {"state": "fresh", "reason": "", **fields}
-
-
-def _earliest_deadline(*values: str) -> str:
-    """The soonest parseable deadline among equivalent spellings, fail-closed.
-
-    ``review_due_at`` and ``stale_after`` are two names for one date, so they
-    normally agree. When something edits only one of them they must not cancel
-    each other out: whichever deadline has already passed decides, so a
-    half-updated record reads as due for review rather than as fresh.
-    """
-    best_value = ""
-    best_time: datetime | None = None
-    for value in values:
-        if not value:
-            continue
-        if not best_value:
-            best_value = value
-        parsed = _parse_utc(value)
-        if parsed is not None and (best_time is None or parsed < best_time):
-            best_time, best_value = parsed, value
-    return best_value
 
 
 def _source_evidence(source_ref: str, *, captured_at: str) -> dict[str, object]:
@@ -4175,35 +3339,6 @@ def _source_evidence_projection(value: Any) -> dict[str, object]:
     if not isinstance(evidence, dict) or not str(evidence.get("sha256", "") or ""):
         return {}
     return {key: str(evidence.get(key, "") or "") for key in ("path", "sha256", "captured_at")}
-
-
-def _local_source_digest(source_ref: str) -> str:
-    """SHA-256 of a local file, or "" when it cannot be digested here."""
-    if not source_ref:
-        return ""
-    try:
-        path = Path(source_ref)
-        if not path.is_absolute() or not path.is_file() or path.stat().st_size > _SOURCE_EVIDENCE_MAX_BYTES:
-            return ""
-        return hashlib.sha256(path.read_bytes()).hexdigest()
-    except (OSError, ValueError):
-        # Unreadable, gone, a directory, or a path this platform rejects. The
-        # caller turns an empty digest into `unreadable`, never into `fresh`.
-        return ""
-
-
-def _source_evidence_state(record: dict[str, Any]) -> str:
-    """"" (no evidence recorded) | unchanged | changed | unreadable."""
-    evidence = record.get("source_evidence")
-    if not isinstance(evidence, dict):
-        return ""
-    recorded = str(evidence.get("sha256", "") or "")
-    if not recorded:
-        return ""
-    current = _local_source_digest(str(evidence.get("path", "") or ""))
-    if not current:
-        return "unreadable"
-    return "unchanged" if current == recorded else "changed"
 
 
 def _project_memory_safety(
@@ -4271,10 +3406,7 @@ def _looks_like_full_transcript(value: str) -> bool:
 # typo rather than an intent, and `created_at + timedelta(days=N)` starts
 # raising OverflowError out of `_days_after` -- which reached the CLI as a raw
 # Python traceback instead of an `omh:` error.
-MAX_RETENTION_DAYS = 36500
 
-_EPISODE_DEFAULT_TTL_DAYS = 30
-_REVIEW_DEFAULT_DAYS = 90
 # The three memory clocks, as product tunables rather than constants baked
 # into one machine's build: the default review cadence, the episode TTL, and
 # the advance-notice window recall packs warn inside (shared by review-due
@@ -4282,11 +3414,6 @@ _REVIEW_DEFAULT_DAYS = 90
 # `memory_policy` -- additive and optional, exactly like the rest of that
 # block -- and an absent or invalid value falls back to the named default
 # while the effective value is always disclosed on the policy payload.
-_MEMORY_CADENCE_DEFAULTS = {
-    "stale_after_days_default": _REVIEW_DEFAULT_DAYS,
-    "episode_ttl_days": _EPISODE_DEFAULT_TTL_DAYS,
-    "due_soon_days": _REVIEW_DUE_SOON_DAYS,
-}
 
 
 def _validated_day_count(days: int | None, *, field: str) -> int | None:
@@ -4416,15 +3543,6 @@ def _days_after(created_at: str, days: int | None) -> str:
     return (base + timedelta(days=int(days))).replace(microsecond=0).isoformat().replace("+00:00", "Z")
 
 
-def _parse_utc(value: str) -> datetime | None:
-    if not value:
-        return None
-    try:
-        return datetime.fromisoformat(value.replace("Z", "+00:00")).astimezone(timezone.utc)
-    except ValueError:
-        return None
-
-
 def _normalize_memory_mode(value: str | None) -> str:
     mode = str(value or "review-first").strip()
     if mode not in PROJECT_MEMORY_MODES:
@@ -4446,35 +3564,6 @@ def _scope_for_project_memory(kind: str, ref: str) -> dict[str, str]:
     if not _SAFE_REF.match(scope["ref"]) or contains_credential_like_material(scope["ref"]):
         raise ValueError("unsafe memory scope ref")
     return scope
-
-
-def _normalize_tags(values: Any, *, redact_sensitive: bool = True) -> list[str]:
-    if not isinstance(values, (list, tuple)):
-        return []
-    tags: list[str] = []
-    seen: set[str] = set()
-    for value in values:
-        raw_tag = unicodedata.normalize("NFC", str(value)).strip()
-        if not raw_tag:
-            continue
-        if raw_tag == "[redacted]":
-            tag = raw_tag
-        elif redact_sensitive and contains_credential_like_material(raw_tag):
-            tag = "[redacted]"
-        else:
-            tag = raw_tag.lower()
-            if not _SAFE_TAG.match(tag):
-                continue
-        if tag not in seen:
-            tags.append(tag)
-            seen.add(tag)
-    return tags[:12]
-
-
-def _string_list(value: Any) -> list[str]:
-    if not isinstance(value, list):
-        return []
-    return [str(item) for item in value if str(item)]
 
 
 def _read_project_memory_candidate(paths: OmhPaths, candidate_id: str) -> dict[str, Any] | None:
@@ -5025,27 +4114,6 @@ def _safe_to_expose_value(key: str, value: Any, item: dict[str, Any]) -> bool:
     return len(text) <= 240
 
 
-def _redacted_metadata_label(value: Any) -> str:
-    text = str(value or "")
-    return "redacted" if contains_credential_like_material(text) else text
-
-
-def _redact_nested_metadata(value: Any) -> Any:
-    if isinstance(value, str):
-        return _redact_admitted_text(value)
-    if isinstance(value, dict):
-        sanitized: dict[str, Any] = {}
-        for index, (key, nested) in enumerate(value.items()):
-            safe_key = _redact_admitted_text(str(key))
-            if safe_key in sanitized:
-                safe_key = f"redacted_{index}"
-            sanitized[safe_key] = _redact_nested_metadata(nested)
-        return sanitized
-    if isinstance(value, (list, tuple)):
-        return [_redact_nested_metadata(nested) for nested in value]
-    return value
-
-
 def _redacted_scope(value: Any) -> dict[str, str]:
     scope = _normalize_scope(value)
     return {
@@ -5077,12 +4145,6 @@ def _redact(value: str) -> str:
     if _looks_sensitive(value):
         return "[redacted]"
     return value[:240]
-
-
-def _redact_admitted_text(value: str) -> str:
-    if contains_credential_like_material(value):
-        return "[redacted]"
-    return value[:500]
 
 
 def _looks_sensitive(value: str) -> bool:
@@ -5531,20 +4593,6 @@ def _handoff_pack_scope(paths: OmhPaths, *, scope_kind: str | None, scope_ref: s
         return _scope(str(scope_kind), str(scope_ref))
     project_root = paths.omh_home.parent
     return _scope("project", project_identity(project_root))
-
-
-def _normalize_scope(value: Any) -> dict[str, str]:
-    if isinstance(value, dict):
-        kind = str(value.get("kind", "project") or "project")
-        ref = str(value.get("ref", "default") or "default")
-        return _scope(kind, ref)
-    if isinstance(value, str) and value:
-        return _scope("project", value)
-    return _scope("project", "default")
-
-
-def _scope(kind: str, ref: str) -> dict[str, str]:
-    return {"kind": kind, "ref": ref}
 
 
 def _source_refs(inspection: dict[str, Any]) -> list[dict[str, object]]:
